Clean up %base-infix.r
The historically dodgy %base-infix.r file has always had the problem
that it needs to define operations that don't have SET-WORD! forms
in R3-Alpha, such as `<:`.  It would need SET-WORD!s, and at the top
level "scope", for them to get scanned and added to LIB.

Because it was complicated to understand and debug what was going on,
the file wound up being somewhat convoluted in Ren-C, with extended
"talking-out-loud" about the nature of the problems as they were hit.

This tightens up the comments and makes it as "table-driven" as it can
be at the moment.  R3-ALPHA-QUOTE is introduced to help annotate points
which should not be changed "to be prettier" as it might break LOAD
compatibility with R3-Alpha for bootstrap.