PARAM_CLASS_NORMAL left comparators and short-circuit boolean ops
This is a major and powerful expressive change, which takes advantage
of the design of "normal" parameters in left-hand enfix positions,
as well as the "DON'T" mechanic which can skip over units of DO/NEXT
to find the next position without evaluating them:

    >> x: 10

    >> if block? x and length of x = 1 + 2 [
        print "x is a block of length 3"
    ] else [
        print "x is either not a block, or not length 3"
    ]
    x is either not a block, or not length 3

    >> x: [powered by MAGIC!]

    >> if block? x and length of x = 1 + 2 [
        print "x is a block of length 3"
    ] else [
        print "x is either not a block, or not length 3"
    ]
    x is a block of length 3

Among many things worth noting in the above example, there is no error
given in the first case resulting from trying to take `length of 10`.
That is because AND is a short-circuiting operator, which acquires
one fixed arg and then optionally will either DO or DON'T a second
variadic argument.

By keeping math operators "left #tight" and moving the boolean ops
and comparators to be "left normal", a balance is struck in the
expression logic.  So it is now possible to write:

    >> if 1 + 2 = 4 - 1 [
        print "both sides are 3"
    ]
    both sides are 3

Some things change, but represent relatively minor collateral damage--
and possibly could be seen as improvements:

    x: true
    y: false
    >> if not x = y [
       print "acts as `if (not x) = y` vs. `if not (x = y)`"
    ]
    acts as `if (not x) = y` vs. `if not (x = y)`

It is worth pointing out that most languages (C included) take such
an interpretation, with logical negation being "lower precedence" than
comparison operators.  And there are options in Rebol--more so than
other languages--e.g. `unless x = y`, in addition to `if x <> y` and
`if not (x = y)`.

Also worth pointing out is that what is going on in the examples above
*is not precedence*.  It is merely allowing some enfixed operators to
choose to have their left-hand arguments gathered w/"full evaluation",
somewhat parallel to how right hand arguments default to a full DO/NEXT
worth of work (as opposed to being quoted, #tight, etc)