Advancements in FILE OF and LINE OF, including for FUNCTION!
This makes reflectors for FILE and LINE, so that they can be used with
OF.  (Hence FILE OF FOO, LINE OF FOO.)  They are specialized as FILE-OF
and LINE-OF.  The ability to get file and line numbers from a function
is also added...though there still needs to be some work on deciding
what the best heuristic is.

It now uses knowledge of the file and line where the evaluator is
processing to tag newly created arrays.  For an operation like a
COMPOSE or REDUCE, this will indicate the array being processed...
because an evaluative frame has been set up on that array.

COPY also tries to use the file and line of the original material.  So:

    foo: [a b c] ;-- line 1
    ...
    ...
    bar: copy [a b c]
    print line of bar ;-- will be 1

These features are still very new, and still has no accounting for the
number of lines in the headers of files...so it counts lines after
the header ends.