Coalesce error and debugger "near" reporting
There were two different bits of code that would build the "near" information...one used in backtraces and one used in error objects.  This brings the best features from both together, and also changes the symbol for where the execution point is to `~~` instead of `??`.

The commit also gets rid of the idea of file and line numbers from C being put into Rebol errors in the debug build.  The increasing utility of actually reporting a rough location in Rebol files is much more important, so that behavior is promoted to be used in all builds.

Fixes a bug where file and line numbers weren't being put on the topmost array in the scanner.  Line numbers still do not account for the header.