Permit GROUP! at head of path, BLOCK!s in paths, etc
R3-Alpha interpreted `(a)/b` as `(a) /b`, so a GROUP! followed by a REFINEMENT!.  It is hard to explain why someone who wanted that wouldn't have put the space in there in the first place.  But certainly most people writing that intend to get a GROUP! at the head of a PATH!.  This commit adds that, as well as support for BLOCK!s in PATH!.

While there are uses in DO for such a thing, there are also dialecting applications.

The method used is to specifically wait to look for slashes until after an entire token has been processed.  Then the resulting value that was pushed to the stack is delegated to the PATH!'s stack by adjusting the data stack pointer by one.

Due to the method by which GET-PATH! is constructed, it is not currently possible to put a BLOCK! or a GROUP! at the head of a GET-PATH!.  SET-PATH!s which end in GROUP! continue to work, and this applies to BLOCK! as well.