General Path Tightening, REBPVS replaced with REBFRM
How R3-Alpha implemented paths was very haphazard and circuitous, based on a set of somewhat confusing conditions about when a path dispatch had located a variable which could be written to vs. when it had created a generative result.  Trying to maintain this code in the face of other system changes was difficult, due to its crypticness and that it had been extended by third parties to do needed features in an ad-hoc way.

Additionally, the original path implementation did not account for GROUP!s and arbitrary code execution during path evaluation, which raised many issues about GC safety (among other things).  Among those issues were that the state structure for the path evaluation was not visible to the GC.

Ren-C added some guards when the problem was noticed, but it seemed a better way of dealing with it would be if instead of a private REBPVS structure, PATH! mechanics used ordinary frames and fetching mechanics.  This could keep more common code paths exercised.

This makes some starting changes which do make what was intended to be going on more clear.  As part of that, it improves the error messages and conditions--in particular to be able to know when a SET-PATH! operation is writing to a temporary location, and hence not having any user-observable effect.

This hopefully can be the groundwork for coming up with a more elegant way of dealing with "sub-addressing" into data that isn't Rebol values.  For example, a GOB! doesn't use a REBVAL pair to store its coordinates, it compresses them in a C struct... yet it expects to be able to say `gob/coordinate/x: 10`...which means the idea of accruing state as pointers to REBVALs is insufficient for the updating of immediate values which are addressed in more than two path steps from a value.

It also makes various path evaluation contexts not allow code execution, e.g. GET and SET are willing to work with paths that don't have GROUP! but give errors indicating REDUCE and EVAL as alternatives.

Also fixes an issue with enfixing via PATH!.