Move all C++ checks under CPLUSPLUS_11 macro
Initially there was some aim to make a plain C++98 compiler to fall
back on some reduced amount of checking where possible of the C code.
That is kind of a silly goal, so it's better to just make a macro
saying that all the checks are for C++11 or later.  That way they can
just say:

    #ifdef CPLUSPLUS_11

instead of:

    #if defined(__cplusplus) && __cplusplus >= 201103L

Though this change had been considered for a while, it was finally
motivated by realizing that Visual Studio has not bumped the
__cplusplus version constant from the 1997 string, even in the 2017
version.  They also don't let you redefine __cplusplus to any other
value (the one part of the standard they choose to be compliant with?)

By having it be a define, it's possible to pass it on the command line
and get the C++11 checking running in MSVC.