STRUCT!/GOB! exemption from end-of-path write assertion
The original concept behind PATH! dispatch was that it would move one
step at a time, producing a new value.  If it was a set-path and
reached the end, it would have access to the value to assign and
perform the assignment.

GOB! and STRUCT! have an issue which made them reach outside the
traditional one-step-at-a-time path dispatch and do two dispatches in
a single call.  e.g. `gob/position: 10x20` could work normally, but
something like `gob/position/x: 10` could not return something from
the `gob/position` step without forgetting which gob and which
position you wanted to write to.  No actual PAIR! was available for
the position as it was compacted C struct data, so `gob/position` was
dynamically constructed...and writing to that pair's X value would
just lose the information vs. writing back to the GOB!.

Coming up with a better solution has been on the agenda, but the path
code has been undergoing some amount of improvement to make a better
design easier to write and test.  This eases up an aggressive assert
that the path code wasn't doing what the STRUCT! and GOB! code did,
to get things like the %gtk.r FFI demo working again.