Preserve enfix status of keys in objects created with COPY
This introduces a specialized kind of movement for variable slots,
Move_Var()...which unlike typical value movement, does not clear the
enfix state of the moved value.  Some special cases did this before,
but this is an attempted survey of all relevant cases to use a special
operation to signal the intent.

Updates README.md, and fixes assert on unknown reflectors.