Don't show "anonymous" filename in error messages
When code is executed via C at the top level with no file or line to
potentially blame for it, the file in an error was getting set to
the "anonymous" symbol.  This translates anonymous filenames into a
BLANK! in the error message, and then doesn't show that blank.

Note that a line number may still be available, e.g. for multi-line
console input...even when there is no associated file.