Workaround __builtin__memcpy_check, bring back %f-deci.c inlining
A warning that triggers under some circumstances was seemingly unhappy
about passing a product of signed REBINT and a REBCNT to a memcpy
call in %f-deci.c.  There was only one invocation at a callsite which
passed 3, and it's not clear why the parameter would need to be signed
anyway.  Changing to unsigned appears to make the warning go away.

Also updates the header comment in %f-deci.c - and since it had an
outdated reference to removing the INLINE annotation, this just goes
and puts back an ordinary `inline` where the old INLINEs were in
R3-Alpha.  Since this involves modifying function prototypes, this
goes ahead and makes all the function prototypes in the file match
the current conventions.