PROBE() passes through value in C++, add BREAK_ON_TICK macro
This adds two debugging conveniences.  One is that in the C++ build,
type templating is used to make it so that PROBE() can return the
exact type, value, and cv-qualifiers it gets in...like Rebol's PROBE.

This means it can be used e.g. as

    return PROBE(Get_Some_Mutable_Series(...));

...or...

    return PROBE(Get_Some_Const_Value(...));

The other addition is something which probably should have been added
much sooner.  Previously you could break on a tick, but it would always
put you in the evaluator loop at the point where the tick was bumped.
If the location you wanted to debug was somewhere else, you'd have to
go put a breakpoint there...but then worry about disabling that
breakpoint on the next debug run and re-enabling it.

This sidesteps all that just by having a BREAK_ON_TICK() macro you can
put with the appropriate number right at the spot you want the debugger
to land.  Though it requires recompilation, it can be well worth it
for complex debugging situations.