Only BLOCK!/FUNCTION! branches for conditionals (IF, CASE, etc.)
This commit backpedals on a change adopted early after R3's open
sourcing, which allowed simple values to be used as branches in a
conditional.  Hence instead of writing:

     x: if 1 < 2 ["one is less than two"]

You could simply write:

     x: if 1 < 2 "one is less than two"

This was in the spirit of Rebol's reluctance to require boilerplate
unless it was absolutely necessary, such as not having the condition
of the IF be in blocks.  The concept wasn't necessarily that people
would stop using the delimiters, but that they might feel other choices
would give them "more bang for the buck" for each character, perhaps
in this case preferring:

     x: if (1 < 2) "one is less than two"

This had some drawbacks, for instance that writing `if condition body`
would look up the variable corresponding to body even if the condition
was false, and it didn't use the fetched value.  However, that was
already the case if `body` is a BLOCK!, so it seemed like no truly
new issues were being introduced.

Time and experience revealed that it actually did introduce somewhat of
an obscuring element.  BLOCK!s (and FUNCTION!s) would execute code
during a kind of "double-evaluation".  Other types would evaluate
the parameter at the callsite, but then not be further processed inside
of the conditional.  For a reader of code (or a writer) it was easy
to make mistakes or be confused about what was going on.

Attempts to add protections to make sure that only literals were being
used in the branch slots for non-BLOCK! and non-FUNCTION! types made
some progress.  However, that impeded abstraction...where code wishing
to pass through an arbitrary branch would have to worry that the act
of passing it through meant transferring a literal via a variable, thus
being against the rules.

In the meantime, conditional operators were introduced that never
evaluated their "branches", but passed through the values directly.
These were ??, !!, and CHOOSE.  Having these as an option made it less
necessary to have branches that weren't blocks, and also addressed
the problem of how to literally return a block (besides [[nesting it]])

This change affects all branches, hence it causes a backwards
incompatibility with R3-Alpha's CASE, which was previously tolerant of
non-BLOCK! branches, so you could write `case [x 1 y 2]`.  One can
either change that to use blocks, or use the CHOOSE operation, which
returns the chosen item literally...even if it's a block.