WAS operator for getting last value while setting a new one
This introduces the WAS operator, which fetches the prior value of a
SET-WORD! or SET-PATH! and evaluates to that, but also performs it as
an assignment.

    >> x: 10

    >> was x: 20
    == 10

    >> x
    == 20

It works based on variadic lookahead, which means it doesn't actually
consume the SET-WORD! as a parameter.  This means it is compatible with
operations that quote backwards to find the SET-WORD!, for example
the DEFAULT operator.

    >> x: _

    >> was x: default [20]
    == _

    >> x
    == 20

Also includes some stray changes to %sys-load.r which make it use the
short circuit AND.