Make SPECIALIZE'd revoked refinements specialize out arguments
If you specialized a function such that a refinement was removed from
the spec by setting it to false, this didn't automatically remove the
arguments to that refinement.  Instead, the argument would appear to
be a normal non-refinement argument.  There was no way to address this,
since the only way to get rid of an argument would be to provide it
a value--and since the refinement was false, any value (even blank)
put into the unused argument would cause an error when the function
was called.

This adjusts the code to remove the arguments from disabled refinements
in the specialization.  It also does a bit more error checking at the
time of specialization, and reuses the errors which the evaluator
would call if such a faulty specialization were used.

Includes a patch to HELP where it was using a non-block branch in an
ELSE, which was causing an error.

Also changes FIND? to use the clearer SOMETHING? instead of TO-LOGIC?
for its chained-to test.