Fold AND~/OR~/XOR~ in INTERSECT/UNION/DIFFERENCE, AND*=>AND+
This folds the prefix form of "bitwise" math into the corresponding
set operations, and then makes AND+/OR+/XOR+ the enfixed versions of
those operations.  This broadens the applicable types that can be used
with the infix operator:

    >> "ade" or+ "bcf"
    == "adebcf"

It also gets rid of the not-very-well-named prefix forms of the math
operation.

This change was proposed long ago to accompany the switch of plain
AND and OR to be "conditional" boolean logic operations.  However, two
details were glossed over.  One is that what should happen on a
BINARY! is slightly ambiguous, as it might be desirable to treat it
as a set of bytes vs. as something to operate on bitwise.  The other
issue is that DIFFERENCE on DATE! is an outlier in when the behavior
is conceptually an XOR, and is more a play on the word "difference"
than an actual fit for the operation.

This commit punts on both issues for the moment, preserving the legacy
behavior on both counts.