Fix whitespace issues found by updated source-analysis.reb
The source analysis script which looks for badly formed whitespace (and
can be hooked for other lint-y things) was previously only checking the
C code in the main source directory.  It was recently updated to scan
extension code, mezzanine code, tests, and other places.

This fixes the issues the broadened scan found.  It also updates the
script which scraped "libuuid" off of GitHub to translate that code's
tabs into spaces, as it was already doing some amount of modification
to the files to make them work with RIGOROUS=yes and C++ builds.