Patches for RIGOROUS: yes in MSVC w/C++17, plain C
This corrects an issue where a void pointer was used in post-vista
builds, where the C++ type checking in Visual Studio expected a pointer
typed as CONSOLE_READCONSOLE_CONTROL.

It also throws in the towel about trying to worry about the compiler
warning discerning `foo()` and `foo(void)`.  In C++ these mean the
same thing, but in C they are different...where the former is not
specifying the interface of the function.  However, Microsoft in their
own header files (Windows.h, ODBC) break the rules in various places
to the point where trying to enforce the warning just litters things.
Disable the warning in MSVC builds, and hope the Linux build (and make
prep) catches it well enough in places where it matters.

Juggles include orderings around in the Crypt module so that a macro
redefinition of min() and max() is worked around.  Once again this
comes down to Visual studio being boneheaded, and defining min and max
as macros in stdlib.h, which is a terrible idea:

https://stackoverflow.com/a/4234022