MAYBE+REALLY => arity 1, MATCH+ENSURE => arity 2
This addresses the question of what the "anti-default" should be called
by taking the name MAYBE:

    >> x: maybe 1 + 2
    == 3

    >> x: maybe blank
    == 3

    >> x: maybe ()
    == 3

    >> x: maybe* blank ;-- * means count _ as value
    == _

    >> x: maybe [1 + 2] ;-- not what you meant
    ** Error: Literal block! used w/MAYBE, use () if intentional

All in all, it's a very good name for what it does.  It's a tool for
*maybe* doing an assignment, if the right hand side is "something"...
where something usually means not-void and not-blank, but can be
fallen back upon to mean just not-void.

A parallel function called REALLY operates in much the same way, but
instead of not doing the assignment it will fail.

    >> x: really ()
    ** Script Error: really requires value argument to not be void

    >> x: really blank
    ** Error: REALLY received a BLANK! (use REALLY* if this is ok)

The names were too good for the purpose to pass up.  But this meant
having to re-shuffle the names for the arity-2 functions, which instead
of assuming blanks and voids are what you're filtering, allow you to
filter on typesets or by calling a function.

Those were given the names MATCH and ENSURE.

    >> match [integer! string!] 3
    == 3

    >> match [integer! string!] <tag>
    == _

    >> match :even? 3
    == _

    >> match :odd? 3
    == 3

    >> ensure [integer! string!] <tag>
    ** Error: ENSURE did not expect argument of type tag!

MATCH is a noun-like word which might be desired for variable names,
and has been used previously in the QuarterMaster web framework.  But
it seems to work reasonably well here.  And the revised words seem to
go along better with their arity.