Kill IF?/FIND?/SELECT?/WHILE?/ALL?/ANY?/CASE?/SWITCH?/PARSE?...
When the ability to specialize and chain functions was introduced in
Ren-C, void-based protocols did not exist (such as those now used by
ELSE and THEN, as well as the /ONLY variations of conditionals).  To
tell whether a conditional had taken a branch, it was necessary for the
native to expose a refinement (called /?) that would return a LOGIC!.

Hence specialization and chaining were used to create a group of
functions that ended in ?.  Unlike most functions with that naming
convention that were simple tests taking one parameter, these would
take as many parameters as the function they specialized, but merely
come back with a LOGIC! answer instead of their usual result.

Most uses of these functions were somewhat gratuitous...not to clamp
values to LOGIC!, but to ostensibly inform the reader that the context
the call was in would only have a logic result used.  Yet the plain
version of the functions which returned a BLANK! would suffice.

Some of the lack of need for the functions came with reaching some
compromises on things like SELECT returning BLANK! when a value was
not found--even though that could be conflated with finding a value
that was actually a blank.  The compromise of having a "core" version
of (e.g. SELECT*) which would use void to signal missing values made
the variants like SELECT? less necessary.

Also, the introduction of a positive analogue to NOT, called DID,
looks relatively good without a disruptive ? for those cases that
really need it.