Temporary simple string evaluation in rebDo()
For starters, the rebDo() API function was based on the C va_list DO
which is used by C code to invoke Rebol expressions.  However, that
was not designed to handle UTF-8 strings initially...the feature of
being able to splice values (and follow other instructions) is under
development.

As an interim solution before the full functionality is available,
this special-cases passing a UTF-8 string and an END marker to rebDo.
The corresponding code is bound to user (with expansion for SET-WORD!)
and then as a fallback bound to lib (with no expansion).

The following code, for instance, would PROBE a result of 7:

    REBVAL *seven = rebDo("foo: :add | foo 3 4", END);
    PROBE(seven);
    rebFree(seven);