Fix aggressive assert in Collect_Unique_Words
The COLLECT-WORDS native called into Collect_Unique_Words, and one
parameter would be a list of words to ignore in the collection.  Since
the user could pass in any array they wanted, the possibility existed
for this to contain duplicate words.  This caused an assert.

This commit does a slightly overkill remedy in the debug build, which
has duplicates count how many duplicates are added and then doesn't
remove until the count goes to 1.  But it helps exercise the binding
code a little bit.

Adds minimal test for the case that triggered the problem.