Repurpose ALSO, kill after, add THEN, swap ELIDE/COMMENT behavior
The ELIDE routine was originally designed as an attempt to create a
"truly invisible" comment.  If given a BLOCK!, it would act like
COMMENT, and if given a GROUP!, it would execute code.  Both could
be done without interfering with the flow of surrounding evaluation.

The reason COMMENT was not done this way is because the definition of
a truly invisible operator wound up being a little weird.  It would
have to be--in effect--an enfix tight operation which was able to
"beam" the state after its parameters were gathered back to be its
return result.  This meant when ELIDE executed code, the moment of
execution made it appear fairly erratic...when most operations that
wanted to be "invisible" would prefer to sacrifice true invisibility
in order to execute code at a more reasonable moment.

ELIDE's moment of execution was so unlikely to be useful that its
ability to execute code was mostly unusable.  It was really only useful
when making a truly invisible comment...a role that seemingly would
be left better to COMMENT.  That would free up ELIDE to be a more tame
invisible operator that would appear to run in sequence.

Making ELIDE more accessible to the average user meant those writing
invisibles could copy from it, and not worry about COMMENT's more
unusual definition...which was really only useful for commenting.  This
flipped the educational bias of the invisibles--and with a usable
ELIDE, the ALSO function was no longer necessary.  One could just
elide any calculation in a more freeform way:

    also a b => a elide b

This meant that AFTER and ALSO could be killed, but the desirable name
ALSO could be used more effectively.  It was taken to mean what the
not-widely-known-about THEN operator had meant, allowing THEN to be
an infix form of IF.  That created parity of THEN/ELSE with ??/!!.