Remove /ONLY from PRINT, in favor of WRITE-STDOUT
Attempts to try and shoehorn all the potential desires of PRINT into
one function seem to be a bad idea.  At one point, /ONLY was conceived
as a switch to tell PRINT to not put in implicit spaces to a block, and
also not to add a terminal newline...leaving people to do their own
custom formatting.  But there's not really a good reason to call that
"PRINT" anymore...one can write directly to STDOUT, the way PRINT does
now in its implementation.  This is more appropriate than asking PRINT
to "not be print"

In historical Rebol, the PRIN native was similar to PRINT but did not
output a newline.  Writing a string to STDOUT will achieve the same
effect in Ren-C, but the standard output does not know how to interpret
a BLOCK!.  Hence if one wishes to use a specification dialect but not
output a newline, that must be called explicitly...such as with:

     write-stdout unspaced [...]

While some code naturally requires the ability to output without a
newline (such as a prompt in the console, where input is to be read on
the same line), most could be solved using COLLECT.