Rename rebFree() to rebRelease()
rebFree() suggests a synchronous and guaranteed freeing of a resource,
which isn't really what's happening when you call the routine.  For
instance, rebFree(some_block) doesn't actually free the block, it just
releases the hold the API handle has on it.  Other references to that
block--from other API handles or just being kept live in the transitive
graph of source--will keep it alive.

Also, it may be that even API handles themselves aren't freed.  There
might be some kind of reference counting strategy by which they can
be rebAddRef()'d and only freed when the last reference goes away.

This changes the name to the more accurate rebRelease()