Introduce ME operator, kill off ++, --, **, and //
Debates about whether to take `//` or `--` for comment based on their
"line-y" looks triggered the question of how these symbol-based items
might be given to purposes besides rarely-used math operations.  It
seems much more Rebol-like to have a test dialect be able to do the
likes of the following when saying the following tests relate to issue
#1020

    -- #1020 --
    [x > 1]
    [x < 2]

Freeing up the "drawing-ish" operator repetitions for users to assign
for such users makes more sense than hardcoding them for math.  So
the idea of a generic ME operator came up, where:

    >> some-var: 20

    >> some-var: me + 1 * 10

    >> print some-var
    210

Being able to write `some-var: me + 10` isn't as "simple" <ahem> as:

* making ME a backwards quote operator that fetches some-var
* quoting its next argument (e.g. +) for word looking up to a function
* making the next argument variadic, and normal-enfix TAKE-ing it
* APPLYing the quoted function on those two values
* setting the left set-word (e.g. some-var:) to the result

The problem with that strategy is that the parameter conventions of +
matter.  Removing it from the evaluator and taking matters into one's
own hands means one must reproduce the evaluator's logic--and that
means it will probably be done poorly.  It's clearly not as sensible as
having some way of slipping the value of some-var into the flow of
normal evaluation.

This commit uses a small trick to implement this specific feature, but
it should be generalized in the future.  However, it is good enough
that ++ and -- can be retaken, and it takes // out (which didn't match
that pattern anyway, and makes for a poor WORD!) so that it may be
repurposed as a comment.  ** is eliminated just for good measure.