Fix incorrect usage of !! in console message
The !! operator is not to be used when not in conjunction with a ??,
unless you really know what you are doing.  The code:

    spaced ["(CONSOLE" (proto-skin/was-updated !! {not}) "updated)"]

Doesn't do what the person who put the !! there (me) intended.  It
looks for a left-hand *branch*, not a left-hand *condition*.  Since
WAS-UPDATED is a LOGIC!, it is always a "branch taken" signal because
it is not void.  So this either has to be written as:

    spaced [
        "(CONSOLE" (not proto-skin/was-updated ?? {not}) "updated)"
    ]

or

    spaced [
        "(CONSOLE" (proto-skin/was-updated ?? () !! {not}) "updated)"
    ]

This chooses the former as being more clear.

Also switches to use error/id instead of error/code numerically for
scan errors in the console, and introduces a test usage of ALSO and a
lambda expression to reduce some code size a small amount.