MY operator for prefix functions (complement to ME for enfix)
This commit adds an implementation for MY that allows a SET-WORD!'s
or SET-PATH!'s existing value to be slipped in the first argument to a
function whose result is then stored back into the variable. e.g.

    >> block: copy [a b c d]

    >> block: my next

    >> probe block
    [b c d]

    >> block: my append [1 2 3]

    >> probe block
    [a b c d 1 2 3] // append returns head of series

Unfortunately the trickery involved does not work for PATH!s at this
time (e.g. append/only), only for WORD!s.  The reason is because it
is actually pretending the right hand argument is an enfix function
even though it is not, and running through the same machinery that
fills the first frame argument of enfix functions while doing the
rest of an evaluation normally.

(A less dodgy technique for doing the same thing should be possible to
design in the future--e.g. this is not something beyond the reasonable
capability of the evaluator to do "correctly".)