Fix assert when data stack overflows occur
The data stack position (DSP) is incremented prior to testing to see
if it bumped into an END marker.  This is a quick way of finding the
limit of the data stack without needing to check the length, and only
examining a cell which is about to be overwritten anyway.  However,
if an expansion failed, this left the pointer on an END marker...
which breaks the invariant that a pushed data stack value can't be
an END.

This simply drops the data stack position by one prior to raising a
failure.

It also renames Trap_Stack_Overflow() to Fail_Stack_Overflow(), as
calling the moment of failure a "trap" is residual R3-Alpha terminology
that is no longer used.  Trapping is instead used to refer to the code
which ultimately catches the failure.