Fix 3 failing CALL/OUTPUT tests on Windows
A test of CALL was added, which used a hardcoded path of %../make/r3
to find a Rebol executable and invoke it.  This had two problems:
on Windows it meant that the .EXE suffix was missing, and also it was
using the "block-style" mode of invoking CALL (e.g. supplying each
individual argv item), which Windows does not support at this time.
(Its process invocation does not take an argv array, so this would
have to be faked somehow.)

This addresses both problems by using system/options/boot to run using
the same executable that is running the tests, and then by building
a STRING! using UNSPACED and translating the path to Rebol to a
local path via FILE-TO-LOCAL.

Includes deletion of an obsolete comment, and a couple of uses of ??
and !! added to %prot-tls that were idly