Fix memory dereference on CALL error, fix bad startup countdown
CALL was freeing the memory holding an integer error code return result
prior to when that memory was used.  This moves the free to after the
usage.

It also corrects the use of a GROUP! in a block being COMPOSE'd; this
meant that the context of the code in the group did not match the
surrounding code where the variables were supposed to be bound.

Additionally, it fixes the broken fixes for the CALL tests from the
last commit--which worked on Windows but would error on Linux.