Make infix AND and OR more flexible on return types, NAY, !?
The search to find some kind of ELSE-like behavior to use with things
like ANY and ALL wound up on the option of utilizing infix OR.  Since
OR is short-circuit of its second clause in a GROUP!, it has the option
of acting somewhat like an ELSE.

(Several languages, such as perl, use OR in this way.)

By tweaking the return result matrix, OR can remain a LOGIC! operation
if both its inputs are LOGIC!...but if it encounters any non-LOGIC!
values then it will return either the last non-LOGIC! value it
evaluates (for a truthy result) or a BLANK!.

This hybridization is also applied to enfix AND, which becomes a sort
of "logic-on-the-left ALSO" (as opposed to checking for voidness on
the left argument).

These are submitted for further testing to see how they work out in
practice.  Also adds NAY and !? operators, as a way of testing for
logic on the left but only acting if the logic is false (e.g. enfix
UNLESS operations).