Remove ENSURE*, fix HELP for specializations, MAYBE return void
`ENSURE*` doesn't make any sense as a concept.  The /ONLY for
EITHER-TEST instructs it to return voids "as-is" when the branch is
run.  But all ENSURE's branch ever does is fail.  If voids are legal
in the test they will be returned anyway, e.g. *(`ensure :void? ()`)*.
Hence ENSURE is all that is needed.

There was a problem in getting HELP on specializations where it would
COPY help information from the parent function.  The COPY was not
needed, and would fail on non-copyable datatypes (e.g. LOGIC!).  This
gets rid of the COPY and makes use of MAYBE.

That MAYBE usage pointed out that although the clause passed in may
not be void, the overall return result might be void if the variable
being maybe-assigned was void to start with.  Make the return `<opt>`