Put just the WORDS-OF a FUNCTION! in its console molding
The default MOLD of a value in the console for output gave too much
information for functions, including its entire source code and
documentation of its parameters.  Rebol2, R3-Alpha, and Red print
nothing when you make a FUNCTION! value.

Since the console is usermode now in Ren-C, it's easy to implement
a happy medium.  This just molds out the function's WORDS-OF in a
construction looking syntax:

    >> foo: function [x {The X} y {The Y} /z {Refine}] [
    [\  print "Function output in console was too much"
    [\  print "Why didn't we shorten it sooner?"
    [\  ]
    == #[function! [x y /z] ...]