Load extensions earlier in boot for GET-ENV console use
GET-ENV had been moved out of the core into the Process extension.  But
it was in the usermode console code as part of startup to detect the
"HOME" directory.  This means that the home directory detection would
not be available if Process was disabled in the build configuration.
But even if it was enabled, it wouldn't work because the extensions
were not loaded by the time the use occurred...and GET-ENV had no
binding in the copy of LIB that the console runs in.

An ATTEMPT masked the fact that none of this works.  This takes out
that ATTEMPT, instead wrapping the search for GET-ENV in the Process
module in an ATTEMPT.  If --verbose is used will tell you why it
couldn't find the home dir, if it can't.

It also moves the load of extensions earlier in the boot.  This may
not be optimal, as there could be some command-line options that would
influence the extension startup code.  Currently that doesn't exist,
but the mechanic should be reviewed when it does.