Stricter rule for infix AND/OR: left hand LOGIC! means logic result
Looking at potential pitfalls of the use of AND and OR as a surrogate
for ELSE, the short circuit was creating a somewhat unpredictable
nature for the return result.  While the left hand side might be a
LOGIC! variable and seem to return LOGIC! when it was FALSE and did
not run the short circuit branch, running that branch might throw in
something a BLANK! and then coerce the result to BLANK!.

This change tries pinning the return result on the first argument.  If
it is LOGIC!, then regardless of the type of the short circuit branch
(which may or may not be run) then the result will be a LOGIC!.  This
should provide a more comfortable invariant for those wishing to
use AND and OR to assign to boolean variables.

Includes a test usage of OR in the host-console and makes the DIE
function for exiting startup accept BLOCK! as well as STRING!.